Move employer list/detail into Employer folder
