+++ src/Sfa.Tl.Find.Provider.Web/Pages/Employer/EmployerDetails.cshtml
@@ -6,7 +6,7 @@
     ViewData["Title"] = @Model.EmployerInterest?.OrganisationName + " | Employer details";
 }
 @section breadcrumbs {
-    <a asp-page="/EmployerList" class="govuk-back-link" id="tl-back">Return to employer list</a>
+    <a asp-page="./EmployerList" class="govuk-back-link" id="tl-back">Return to employer list</a>
 }
 
 <div class="govuk-grid-row">
