+++ src/Sfa.Tl.Find.Provider.Web/Pages/Employer/_administratorEmployerList.cshtml
@@ -1,4 +1,4 @@
-﻿@model Sfa.Tl.Find.Provider.Web.Pages.EmployerListModel
+﻿@model EmployerListModel
 
 @if (Model.EmployerInterestList != null && Model.EmployerInterestList.Any())
 {
@@ -44,7 +44,7 @@
                                 <div class="tl-employers--table--inline-header"><strong>Details expire</strong></div>
                                 <div class="tl-employers--table--inline-cell">@($"{employerInterest.ExpiryDate:d MMM yyyy}")</div>
                             </td>
-                            <td class="govuk-table__cell"><a asp-page="/EmployerDetails" asp-route-id="@employerInterest.Id" id="tl-employer--table--link--@employerInterest.Id" class="govuk-link govuk-link--no-visited-state">View</a></td>
+                            <td class="govuk-table__cell"><a asp-page="./EmployerDetails" asp-route-id="@employerInterest.Id" id="tl-employer--table--link--@employerInterest.Id" class="govuk-link govuk-link--no-visited-state">View</a></td>
                         </tr>
                     }
                 </tbody>
