+++ src/Sfa.Tl.Find.Provider.Web/Program.cs
@@ -58,8 +58,8 @@ builder.Services.Configure<RouteOptions>(option =>
 builder.Services.AddRazorPages(options =>
 {
     //options.Conventions.Add(new PageRouteTransformerConvention(new SlugifyParameterTransformer()));
-    options.Conventions.AddPageRoute("/EmployerList", "/employer-list");
-    options.Conventions.AddPageRoute("/EmployerDetails", "/employer-details");
+    options.Conventions.AddPageRoute("/Employer/EmployerList", "/employer-list");
+    options.Conventions.AddPageRoute("/Employer/EmployerDetails", "/employer-details");
     options.Conventions.AddPageRoute("/Help/AccessibilityStatement", "/accessibility-statement");
     options.Conventions.AddPageRoute("/Help/Cookies", "/cookies");
     options.Conventions.AddPageRoute("/Help/Privacy", "/privacy");
